requirements_new.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.10
django-ckeditor==6.7.1
django-import-export==3.3.7
django-js-asset==2.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
MarkupPy==1.14
numpy==1.24.4
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pillow==10.2.0
pkg_resources==0.0.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
six==1.16.0
sqlparse==0.4.4
tablib==3.5.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
xlrd==2.0.1
xlwt==1.3.0
